BundleSeg : Use SFT for readability + save as float32

+++ src/scilpy/segment/bundleseg.py
@@ -271,7 +271,7 @@ class BundleSeg(object):
         # Neighbors can be refined since the search space is smaller
         t0 = time()
         neighb_streamlines = reconstruct_streamlines_from_memmap(
-            self.memmap_filenames, self.neighb_indices, strs_dtype=np.float16)
+            self.memmap_filenames, self.neighb_indices)
 
         # Typically the neighbors is bigger than the model, so we flip the
         # FSS to be more memory efficient
